+++ maybetype/errors.py
@@ -1,9 +1,8 @@
-class MaybeInstanceWarning(Warning):
-    """Emitted when creating a ``Maybe`` object directly."""
-class ResultInstanceWarning(Warning):
-    """Emitted when creating a ``Result`` object directly."""
-class NothingTypeInitWarning(Warning):
-    """Emitted when calling ``NothingType.__init__()`` for a second time instead of using the ``Nothing`` singleton."""
-
+class MaybeInstanceError(Exception):
+    """Raised when creating a ``Maybe`` object directly."""
+class NothingTypeInitError(Exception):
+    """Raised when calling ``NothingType.__init__()`` for a second time instead of using the ``Nothing`` singleton."""
+class ResultInstanceError(Exception):
+    """Raised when creating a ``Result`` object directly."""
 class ResultUnwrapError(Exception):
     """Raised when calling ``Result.unwrap()`` if the instance is not ``Ok``."""